fix(type mismatch): Fixed type mismatch in process list rendering coordinates
# Project State

## Current Focus
Fixed type mismatch in process list rendering coordinates

## Context
The system monitor example was rendering process list items with a type mismatch between the index and coordinate parameters in the `draw_text_plane` call.

## Completed
- [x] Fixed type mismatch by casting `i` to `u16` in the process list rendering coordinates

## In Progress
- [x] No active work in progress beyond this fix

## Blockers
- None identified

## Next Steps
1. Verify the fix doesn't introduce new rendering issues
2. Consider adding more comprehensive type checking for similar UI components
---
category: fix
scope: type mismatch
---

+++ examples/_apps/system_monitor.rs
@@ -576,7 +576,7 @@ fn main() -> std::io::Result<()> {
             } else {
                 (t.fg, Color::Reset)
             };
-            draw_text_plane(&mut proc_plane, 0, 1 + i, &line, fg, bg, is_selected);
+            draw_text_plane(&mut proc_plane, 0, 1 + i as u16, &line, fg, bg, is_selected);
         }
 
         ctx.add_plane(proc_plane);
